requirements.txt
alabaster==0.7.12
Babel==2.9.1
backports.functools-lru-cache==1.6.4
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
common==0.1.2
Cython==3.0.0a9
docutils==0.17.1
future==0.18.2
Fuzzy==1.2.2
fuzzywuzzy==0.18.0
google-transliteration-api==1.0.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
icu==0.0.1
idna==2.10
imagesize==1.2.0
indian-namematch==1.3.0
indic-nlp-library==0.81
indic-transliteration==2.2.9
indicnlp==0.0.1
indictrans==1.2.3
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
Morfessor==2.0.6
nltk==3.6.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pbr==5.6.0
polyglot==16.7.4
Pygments==2.10.0
pyparsing==2.4.7
pyphonetics==0.5.3
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
regex==2020.10.11
requests==2.26.0
rfc3986==1.5.0
scipy==1.7.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-argparse==0.3.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tqdm==4.62.3
typer==0.3.2
Unidecode==1.3.2
urllib3==1.26.6
